adding more labels to query. added title search for files too
