+++ src/components/QueryView.tsx
@@ -2,11 +2,86 @@ import React, { useState } from 'react';
 import { MdCancel } from 'react-icons/md';
 import { useProjectState } from './ProjectContext';
 import QueryMatchComponent from './QueryMatchComponent';
-
+import {
+  Button,
+  Popover,
+  PopoverTrigger,
+  PopoverContent,
+  PopoverArrow,
+  PopoverBody,
+} from '@chakra-ui/react';
 
 interface QueryViewProps {
   setViewType: (viewType: string) => void;
 }
+
+export const HoverTitle = (props: any) => {
+  const { title, entry, match, setViewType } = props;
+
+  const [{ query }, dispatch] = useProjectState();
+
+  const fileIndex =
+    match.fileType === 'gdoc'
+      ? entry.files.map((m) => m.title).indexOf(match.title)
+      : match['file-index'];
+
+  const [showPopover, setShowPopover] = useState(false);
+
+  const closePopover = () => {
+    setShowPopover(false);
+  };
+
+  return (
+    <>
+      {showPopover ? (
+        <Popover isOpen={showPopover} onClose={closePopover}>
+          <PopoverTrigger>
+            <div style={{ display: 'inline', marginRight: 10 }}>{title}</div>
+          </PopoverTrigger>
+          <PopoverContent bg="white" color="gray">
+            <PopoverArrow bg="white" />
+            <PopoverBody>
+              <Button
+                onClick={() => {
+                  console.log('matches??', match);
+                  setViewType('detail view');
+                  dispatch({
+                    type: 'SELECTED_ARTIFACT',
+                    activity: entry,
+                    artifactIndex: fileIndex,
+                    hopArray: [
+                      {
+                        activity: entry,
+                        artifactUid: entry.files[fileIndex]
+                          ? entry.files[fileIndex].artifact_uid
+                          : null,
+                        hopReason: 'first hop',
+                      },
+                    ],
+                  });
+                  dispatch({
+                    type: 'UPDATE_GO_BACK',
+                    goBackView: 'query',
+                    filterQuery: query.matches.map((m) => m.entry.title),
+                  });
+                }}
+              >
+                See artifact in detail.
+              </Button>
+            </PopoverBody>
+          </PopoverContent>
+        </Popover>
+      ) : (
+        <div
+          style={{ display: 'inline', marginRight: 10 }}
+          onMouseEnter={() => setShowPopover(true)}
+        >
+          {title}
+        </div>
+      )}
+    </>
+  );
+};
 
 const QueryView = (props: QueryViewProps) => {
   const { setViewType } = props;
@@ -43,7 +118,35 @@ const QueryView = (props: QueryViewProps) => {
               <div style={{ fontSize: 18, fontWeight: 700, marginTop: 30 }}>
                 {m.entry.title}
               </div>
-
+              {m.titleMatch.length > 0 && (
+                <div
+                style={{fontWeight: 700}}
+                >Matches in Titles :</div>
+              )}
+              
+              {
+                m.titleMatch.length > 0 && (
+                  m.titleMatch.map((tm, j)=> (
+                    <React.Fragment
+                    key={`title-match-${j}`}
+                    >{
+                      Object.keys(tm).includes('fileTitle') ? 
+                      <div><span
+                      style={{fontWeight: 700}}
+                      >{'File title match: '}</span><HoverTitle title={tm.fileTitle} entry={m} match={tm} setViewType={setViewType}/></div> 
+                      : 
+                      <div></div>
+                    }
+                      
+                    </React.Fragment>
+                  ))
+                )
+              }
+              {m.textMatch.length > 0 && (
+                <div
+                style={{fontWeight: 700}}
+                >Matches in Text files :</div>
+              )}
               {m.textMatch.length > 0 &&
                 m.textMatch.map((tm, j) => (
                   <React.Fragment
@@ -58,7 +161,11 @@ const QueryView = (props: QueryViewProps) => {
                   </React.Fragment>
                
                 ))}
-
+              {m.googMatch.length > 0 && (
+                <div
+                style={{fontWeight: 700}}
+                >Matches in google docs :</div>
+              )}
               {m.googMatch.length > 0 &&
                 m.googMatch.map((gm, j) => (
                   <React.Fragment
